Fix compile error #2

+++ macos-homebrew/include/shared/bus-get-properties.h
@@ -4,6 +4,7 @@
 #include "sd-bus.h"
 
 #include <basic/macro.h>
+#include <linux/types.h>
 
 int bus_property_get_bool(sd_bus *bus, const char *path, const char *interface, const char *property, sd_bus_message *reply, void *userdata, sd_bus_error *error);
 int bus_property_set_bool(sd_bus *bus, const char *path, const char *interface, const char *property, sd_bus_message *value, void *userdata, sd_bus_error *error);
@@ -46,7 +47,8 @@ assert_cc(sizeof(gid_t) == sizeof(uint32_t));
 assert_cc(sizeof(pid_t) == sizeof(uint32_t));
 #define bus_property_get_pid ((sd_bus_property_get_t) NULL)
 
-assert_cc(sizeof(mode_t) == sizeof(uint32_t));
+// assert_cc(sizeof(mode_t) == sizeof(uint32_t));
+assert_cc(sizeof(mode_t) == sizeof(unsigned short));
 #define bus_property_get_mode ((sd_bus_property_get_t) NULL)
 
 int bus_property_get_rlimit(sd_bus *bus, const char *path, const char *interface, const char *property, sd_bus_message *reply, void *userdata, sd_bus_error *error);
